// enums/skill/skill-categories-enum.ts
/**
 * This enum is used to represent the categories of the skills.
 * This enum is used to avoid hardcoding the categories in the code, getting autocompletion support and avoiding typos.
 */
enum SkillCategoriesEnum {
  ArtificialIntelligence = "Artificial Intelligence",
  FullStackWebDevelopment = "Full-Stack Web Development",
  BackEndWebDevelopment = "Back-End Web Development",
  FrontEndWebDevelopment = "Front-End Web Development",
  DatabaseManagement = "Database Management",
  DevOps = "DevOps",
  Testing = "Testing",
  ProgrammingLanguages = "Programming Languages",
  CloudComputing = "Cloud Computing",
  Monitoring = "Monitoring",
  GameDevelopment = "Game Development",
  Mathematics = "Mathematics",
}

export default SkillCategoriesEnum;